requirements.txt
asgiref==3.4.1
Django==3.2.9
django-grappelli==2.15.3
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
